requirements.txt
amqp==2.2.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==2.0
django-allauth==0.35.0
django-celery-beat==1.1.0
django-grappelli==2.10.1
django-social-share==1.1.2
django-tinymce==2.7.0
easy-thumbnails==2.5
ephem==3.7.6.0
future==0.16.0
html5lib==1.0.1
idna==2.6
kombu==4.1.0
Markdown==2.6.11
misaka==2.1.0
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.3
python-twitter==3.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
